+++ org.hl7.fhir.r4/src/main/java/org/hl7/fhir/r4/utils/NPMPackageGenerator.java
@@ -111,7 +111,7 @@ public class NPMPackageGenerator {
     p.remove("name");
     p.addProperty("name", id);
     p.remove("type");
-    p.addProperty("type", PackageType.SUBSET.getCode());    
+    p.addProperty("type", PackageType.CONFORMANCE.getCode());    
     p.remove("title");
     p.addProperty("title", name);    
     return new NPMPackageGenerator(destFile, p);