+++ RELEASE_NOTES.md
@@ -5,3 +5,4 @@
 ## Other code changes
 
 * Ignore META-INF contents from other dependencies while building Validator jar
+* Fix to r5 TestingUtilities to allow NarrativeGenerationTests to pass.